requirements.txt
crewai~=0.30.11
duckduckgo-search~=6.1.0
langchain-community~=0.0.34
langchain~=0.1.16
python-decouple~=3.8
streamlit~=1.34.0
ruff~=0.4.4
